==> 01_bronze_extract_list.ipynb <==
import requests
import json

# URL da PokeAPI para buscar uma lista com até 5000 Pokémon
pokeapi_url = 'https://pokeapi.co/api/v2/pokemon/?limit=5000'

# Faz a requisição GET para obter os dados da API
response = requests.get(pokeapi_url)

# Converte o conteúdo JSON da resposta em um dicionário Python
pokemon_data = response.json()

# Lista todas as chaves presentes no dicionário retornado
pokemon_data_keys = pokemon_data.keys()

# Exibe as chaves disponíveis (por exemplo: 'count', 'next', 'previous', 'results')
print(pokemon_data_keys)

pokemon_data['results']

# Extrai apenas a lista de Pokémon (nome + URL)
pokemon_list = pokemon_data['results']

# Define o caminho no DBFS (Databricks File System) para salvar os dados
output_path = "raw/raw_pokemon_list.json"

# Converte a lista de Pokémon para JSON formatado (legível)
json_formatted_data = json.dumps(pokemon_list, indent=4, ensure_ascii=False)

# Salva o JSON no DBFS, sobrescrevendo se o arquivo já existir
with open(output_path,'w',encoding='utf-8') as file:
    file.write(json_formatted_data)

import os
# Obtém o tamanho do arquivo em bytes
file_size_bytes = os.path.getsize(output_path)

print(f"{file_size_bytes} bytes.")
