==> 02_silver_transform_ids.ipynb <==
import pandas as pd
import re

# Caminho para o arquivo JSON
input_path = "raw/raw_pokemon_list.json"

# Lê o arquivo JSON (multiLine=True não é necessário no pandas, mas precisa estar bem formatado)
df_raw_pokemons = pd.read_json(input_path)


# Função para extrair o ID do Pokémon da URL
def extract_pokemon_id(url):
    match = re.search(r'/pokemon/(\d+)/', url)
    return int(match.group(1)) if match else None

# Aplica a função para extrair os IDs
df_raw_pokemons['pokemon_id'] = df_raw_pokemons['url'].apply(extract_pokemon_id)


# Exibe os primeiros registros e o schema (tipos de dados)
df_raw_pokemons.head()
print(df_raw_pokemons.dtypes)



# Salva em CSV
output_path = "silver/silver_pokemon_ids.csv"
df_raw_pokemons.to_csv(output_path, index=False)

import os
# Obtém o tamanho do arquivo em bytes
file_size_bytes = os.path.getsize(output_path)

print(f"{file_size_bytes} bytes.")
