==> 03_bronze_extract_details.ipynb <==
import requests
import json
import pandas as pd
from concurrent.futures import ThreadPoolExecutor

# URL da API para obter os dados do Pokémon com ID 2 (Ivysaur)
pokemon_url = 'https://pokeapi.co/api/v2/pokemon/2'

# Faz uma requisição GET para a API
response = requests.get(pokemon_url)

# Converte a resposta JSON em um dicionário Python
pokemon_data = response.json()

# Lista as chaves de nível superior do dicionário retornado
top_level_keys = pokemon_data.keys()
  
# Exibe as chaves disponíveis no dicionário da API
print(top_level_keys)


# Lê o CSV salvo anteriormente
df_pokemons = pd.read_csv("silver/silver_pokemon_ids.csv")

# Extrai a coluna de IDs como lista
list_pokemon_ids = df_pokemons['pokemon_id'].tolist()

# Exibe a lista de IDs
print(list_pokemon_ids)

# Função para buscar dados de um Pokémon específico pelo ID
def fetch_pokemon_data(pokemon_id):
    url = f"https://pokeapi.co/api/v2/pokemon/{pokemon_id}"
    try:
        response = requests.get(url, timeout=5)  # Timeout de 5s para evitar travamentos
        response.raise_for_status()              # Lança exceção se o status não for 200
        return response.json()                   # Retorna os dados em JSON
    except requests.RequestException as error:
        print(f"⚠️ Erro ao buscar ID {pokemon_id}: {error}")
        return None

# Lista de IDs a serem consultados (extraída da camada gold)
pokemon_ids = list_pokemon_ids

# Lista para armazenar os dados obtidos com sucesso
pokemon_data_list = []

# Requisições em paralelo usando até 10 threads
with ThreadPoolExecutor(max_workers=10) as executor:
    results = list(executor.map(fetch_pokemon_data, pokemon_ids))

# Filtra e armazena apenas os resultados válidos (sem erro)
pokemon_data_list = [data for data in results if data is not None]

# Caminho no DBFS onde o arquivo JSON será salvo (camada RAW)
output_path = "raw/raw_pokemon_descriptions.json"

# Converte a lista de Pokémon para JSON formatado (legível)
json_formatted_data = json.dumps(pokemon_data_list, indent=4, ensure_ascii=False)

# Salva o JSON no DBFS, sobrescrevendo se o arquivo já existir
with open(output_path,'w',encoding='utf-8') as file:
    file.write(json_formatted_data)

import os
# Obtém o tamanho do arquivo em bytes
file_size_bytes = os.path.getsize(output_path)

print(f"{file_size_bytes} bytes.")
