==> 04_gold_transform_final_data.ipynb <==
import json
import re
import pandas as pd

# Caminho do arquivo JSON
with open('raw/raw_pokemon_descriptions.json', 'r', encoding='utf-8') as file:
    dados = json.load(file)

amostra = dados[0]
amostra['sprites']['front_default']


### Tabela Nomes

# Extraindo os nomes dos pokemons
nomes_pokemons = [item['name'] for item in dados]

# Extraindo as IDs dos pokemons
keys_pokemons = [item['id'] for item in dados]

# Extraindo as Fotos dos pokemons
fotos_pokemons = [item['sprites']['front_default'] for item in dados]


# Junta os IDs e nomes como tuplas
combinado_pokemons = list(zip(keys_pokemons, nomes_pokemons,fotos_pokemons))

# Remove duplicatas mantendo a ordem
unicos_pokemons = list(dict.fromkeys(combinado_pokemons))

# Cria o DataFrame
dim_pokemons = pd.DataFrame(unicos_pokemons,  columns=['keys_pokemons','nomes_pokemons','fotos_pokemons'])

display(dim_pokemons)

### Tabela Movimentos


# Extraindo os golpes
moves = [item['moves'] for item in dados]

# Extraindo os nomes dos golpes
nomes_moves = [[item['move']['name'] for item in sublista] for sublista in moves]

# Extraindo as URLs dos golpes
urls_moves = [[item['move']['url'] for item in sublista] for sublista in moves]

# Extraindo apenas os números das URLs
keys_moves = [[int(re.search(r'/move/(\d+)/', url).group(1)) for url in sublista] for sublista in urls_moves]

# Extrai os IDs dos golpes
keys_moves_nested = [
    [int(re.search(r'/move/(\d+)/', url).group(1)) for url in sublista]
    for sublista in urls_moves
]
# Achata a lista de IDs dos golpes (lista de listas → lista única)
keys_moves_flat = [item for sublist in keys_moves_nested for item in sublist]

# Achata a lista de Nomes dos golpes (lista de listas → lista única)
nomes_moves_flat = [item for sublist in nomes_moves for item in sublist]

# Junta os IDs e nomes como tuplas
combinado_moves = list(zip(keys_moves_flat, nomes_moves_flat))

# Remove duplicatas mantendo a ordem
unicos_moves = list(dict.fromkeys(combinado_moves))

# Cria o DataFrame
dim_movimentos = pd.DataFrame(unicos_moves, columns=['keys_moves', 'nomes_moves'])

display(dim_movimentos)

# Cria o DataFrame inicial com as listas aninhadas
pokemons_movimentos = pd.DataFrame({
    "keys_pokemons": keys_pokemons,
    "keys_moves": keys_moves
})

# Explode a coluna 'keys_moves' (que contém listas) em múltiplas linhas
rel_pokemons_movimentos = pokemons_movimentos.explode("keys_moves")

# Exibe o resultado
display(rel_pokemons_movimentos)

### Tabela Habilidades


# Extraindo as habilidades
habilidades = [item['abilities'] for item in dados]

# Extraindo Nomes das habilidades
nomes_habilidades =[ [item['ability']['name'] for item in sublista] for sublista in habilidades]

# Extraindo as URLs das habilidades
urls_habilidades = [ [item['ability']['url'] for item in sublista] for sublista in habilidades]

# Extrai os IDs das golpes
keys_habilidades = [
    [int(re.search(r'/ability/(\d+)/', url).group(1)) for url in sublista]
    for sublista in urls_habilidades
]

# Achata a lista de IDs das habilidades (lista de listas → lista única)
keys_habilidades_flat = [item for sublist in keys_habilidades for item in sublist]

# Achata a lista de Nomes das habilidades (lista de listas → lista única)
nomes_habilidades_flat = [item for sublist in nomes_habilidades for item in sublist]

# Junta os IDs e nomes como tuplas
combinado_habilidades = list(zip(keys_habilidades_flat, nomes_habilidades_flat))

# Remove duplicatas mantendo a ordem
unicos_habilidades = list(dict.fromkeys(combinado_habilidades))

# Cria o DataFrame
dim_habilidades = pd.DataFrame(unicos_habilidades, columns=['keys_habilidades', 'nomes_habilidades'])

display(dim_habilidades)

# Cria o DataFrame inicial com as listas aninhadas
pokemons_habilidades = pd.DataFrame({
    "keys_pokemons": keys_pokemons,
    "keys_habilidades": keys_habilidades
})

# Explode a coluna 'keys_habilidades' (que contém listas) em múltiplas linhas
rel_pokemons_habilidades = pokemons_habilidades.explode("keys_habilidades")

# Exibe o resultado
display(rel_pokemons_habilidades) 

### Tabela Status

amostra_status = dados[0]['stats']
# Extraindo atributos
status = [item['stats'] for item in dados]

# Extraindo nomes dos atributos para utilizar como nome das colunas
nomes_coluna_status =[item['stat']['name'] for item in amostra_status]

# Tirando caracteres que podem dar conflito na base
nomes_coluna_status = [col.replace("-", "_") for col in nomes_coluna_status] 

nomes_coluna_status

# Extraindo valores dos atributos 
valores_status =[ [item['base_stat'] for item in sublista] for sublista in status]

valores_status

import pandas as pd

# Criar uma lista de dicionários para facilitar a conversão
dados_status_dict = [
    dict(zip(['keys_pokemons'] + nomes_coluna_status, [key] + status))
    for key, status in zip(keys_pokemons, valores_status)
]

# Criar o DataFrame com pandas
dim_status = pd.DataFrame(dados_status_dict)

# Exibir (no Databricks, o display também funciona com pandas)
display(dim_status)



# dim_habilidades # Dimensão de habilidades
# dim_movimentos # Dimensão de movimentos
# dim_pokemons # Dimensão de pokemons
# dim_status  # Dimensão de atributlos
# rel_pokemons_habilidades  # Bridge de habilides
# rel_pokemons_movimentos  # Bridge de movimentos

tabelas_nomes = [
    'dim_habilidades',
    'dim_movimentos',
    'dim_pokemons',
    'dim_status',
    'rel_pokemons_habilidades',
    'rel_pokemons_movimentos'
]

tabelas = [
    dim_habilidades,
    dim_movimentos,
    dim_pokemons,
    dim_status,
    rel_pokemons_habilidades,
    rel_pokemons_movimentos
]
for nome, tabela in zip(tabelas_nomes, tabelas):
    # Caminho onde os arquivos são salvos
    output_path = f'gold/{nome}.csv'
    
    # Escreve como CSV em pasta temporária
    tabela.to_csv(output_path, index=False)



    
    print(f'{output_path} salvo com sucesso!')

# 📥 Download da Tabela

- [📄 dim_habilidades.csv](https://community.cloud.databricks.com/files/pokemon_api/gold/dim_habilidades/dim_habilidades.csv)
- [📄 dim_movimentos.csv](https://community.cloud.databricks.com/files/pokemon_api/gold/dim_movimentos/dim_movimentos.csv)
- [📄 dim_pokemons.csv](https://community.cloud.databricks.com/files/pokemon_api/gold/dim_pokemons/dim_pokemons.csv)
- [📄 dim_status.csv](https://community.cloud.databricks.com/files/pokemon_api/gold/dim_status/dim_status.csv)
- [📄 rel_pokemons_habilidade.csv](https://community.cloud.databricks.com/files/pokemon_api/gold/rel_pokemons_habilidades/rel_pokemons_habilidades.csv)
- [📄 rel_pokemons_movimentos.csv](https://community.cloud.databricks.com/files/pokemon_api/gold/rel_pokemons_movimentos/rel_pokemons_movimentos.csv)
